move the index::Tree into a separate module
