+++ src/index/cli.rs
@@ -1,11 +1,10 @@
-use std::fs;
-use std::io::{self, BufRead, Write};
 use std::path::PathBuf;
 use std::process::{Command, Stdio};
 
 use semver::{Version, VersionReq};
 
-use crate::error::{AlexError, Error};
+use crate::error::Error;
+use crate::index::tree::Tree;
 use crate::index::{CrateVersion, Indexer};
 
 /// The 'command-line' crate index management strategy type.
@@ -22,7 +21,7 @@ impl CommandLineIndex {
     pub fn new<P: Into<PathBuf>>(path: P) -> CommandLineIndex {
         let path = path.into();
         let repo = Repository { path: path.clone() };
-        let tree = Tree { path };
+        let tree = Tree::new(path);
         CommandLineIndex { repo, tree }
     }
 }
@@ -61,134 +60,6 @@ impl Indexer for CommandLineIndex {
         F: FnOnce(&mut CrateVersion),
     {
         self.tree.alter_record(name, version, func)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-struct Tree {
-    path: PathBuf,
-}
-
-impl Tree {
-    fn compute_record_path(&self, name: &str) -> PathBuf {
-        match name.len() {
-            1 => self.path.join("1").join(&name),
-            2 => self.path.join("2").join(&name),
-            3 => self.path.join("3").join(&name[..1]).join(&name),
-            _ => self.path.join(&name[0..2]).join(&name[2..4]).join(&name),
-        }
-    }
-
-    fn match_record(&self, name: &str, req: VersionReq) -> Result<CrateVersion, Error> {
-        let path = self.compute_record_path(name);
-        let file = fs::File::open(path).map_err(|err| match err.kind() {
-            io::ErrorKind::NotFound => Error::from(AlexError::CrateNotFound {
-                name: String::from(name),
-            }),
-            _ => Error::from(err),
-        })?;
-        let found = io::BufReader::new(file)
-            .lines()
-            .map(|line| Some(json::from_str(line.ok()?.as_str()).ok()?))
-            .flat_map(|ret: Option<CrateVersion>| ret.into_iter())
-            .filter(|krate| req.matches(&krate.vers))
-            .max_by(|k1, k2| k1.vers.cmp(&k2.vers));
-        Ok(found.ok_or_else(|| AlexError::CrateNotFound {
-            name: String::from(name),
-        })?)
-    }
-
-    fn all_records(&self, name: &str) -> Result<Vec<CrateVersion>, Error> {
-        let path = self.compute_record_path(name);
-        let reader = io::BufReader::new(fs::File::open(path)?);
-        reader
-            .lines()
-            .map(|line| Ok(json::from_str::<CrateVersion>(line?.as_str())?))
-            .collect()
-    }
-
-    fn latest_record(&self, name: &str) -> Result<CrateVersion, Error> {
-        let records = self.all_records(name)?;
-        Ok(records
-            .into_iter()
-            .max_by(|k1, k2| k1.vers.cmp(&k2.vers))
-            .expect("at least one version should exist"))
-    }
-
-    fn add_record(&self, record: CrateVersion) -> Result<(), Error> {
-        let path = self.compute_record_path(record.name.as_str());
-
-        if let Ok(file) = fs::File::open(&path) {
-            let reader = io::BufReader::new(file);
-            let records = reader
-                .lines()
-                .map(|line| Ok(json::from_str::<CrateVersion>(line?.as_str())?))
-                .collect::<Result<Vec<CrateVersion>, Error>>()?;
-            let latest = records
-                .into_iter()
-                .max_by(|k1, k2| k1.vers.cmp(&k2.vers))
-                .expect("at least one record should exist");
-            if record.vers <= latest.vers {
-                return Err(Error::from(AlexError::VersionTooLow {
-                    krate: record.name,
-                    hosted: latest.vers,
-                    published: record.vers,
-                }));
-            }
-        } else {
-            let parent = path.parent().unwrap();
-            fs::create_dir_all(parent)?;
-        }
-
-        let mut file = fs::OpenOptions::new()
-            .write(true)
-            .append(true)
-            .create(true)
-            .open(path)?;
-        json::to_writer(&mut file, &record)?;
-        writeln!(file)?;
-        file.flush()?;
-
-        Ok(())
-    }
-
-    fn alter_record<F>(&self, name: &str, version: Version, func: F) -> Result<(), Error>
-    where
-        F: FnOnce(&mut CrateVersion),
-    {
-        let path = self.compute_record_path(name);
-        let file = fs::File::open(path.as_path()).map_err(|err| match err.kind() {
-            io::ErrorKind::NotFound => Error::from(AlexError::CrateNotFound {
-                name: String::from(name),
-            }),
-            _ => Error::from(err),
-        })?;
-        let mut krates: Vec<CrateVersion> = {
-            let mut out = Vec::new();
-            for line in io::BufReader::new(file).lines() {
-                let krate = json::from_str(line?.as_str())?;
-                out.push(krate);
-            }
-            out
-        };
-        let found = krates
-            .iter_mut()
-            .find(|krate| krate.vers == version)
-            .ok_or_else(|| {
-                Error::from(AlexError::CrateNotFound {
-                    name: String::from(name),
-                })
-            })?;
-
-        func(found);
-
-        let lines = krates
-            .into_iter()
-            .map(|krate| json::to_string(&krate))
-            .collect::<Result<Vec<String>, _>>()?;
-        fs::write(path.as_path(), lines.join("\n") + "\n")?;
-
-        Ok(())
     }
 }
