+++ src/index/mod.rs
@@ -3,6 +3,7 @@ use semver::{Version, VersionReq};
 /// Index management through `git` shell command invocations.
 pub mod cli;
 mod models;
+mod tree;
 
 /// Index management using [**`libgit2`**][libgit2].
 /// [libgit2]: https://libgit2.org
